Rename s3_stage.S3BucketCreator -> s3_stage.S3Bucket, move to _config to prevent cyclical imports
